pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toolscout"
version = "0.1.0"
description = "Retrieve the most relevant tools for LLM agents from large tool libraries."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "ToolScout Contributors" }]
keywords = ["llm", "agents", "tool-retrieval", "faiss", "sentence-transformers"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "numpy>=1.24",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0.0",
  "openai>=1.30.0",
]

[tool.setuptools.packages.find]
include = ["toolscout*"]
